[issue002a1] Generate and store cryptographically secure flask secret keys
FEATURE(s):
issue002a-auth-python:
  Create the auth application using python
  issue002a1-auth-python-flask:
    Create the auth application using python and flask

FILE(s):
[new]auth/db/erase_secrets_db.sql
[new]auth/db/get_secret_key.sh
[new]auth/db/initialize_secrets_db.sql
[new]auth/db/insert_secret_key.sql
[modified]auth/python/flask/auth/auth/auth.py
[modified]auth/python/flask/auth/run-app.sh
[new]auth/python/flask/auth/tools/.gitignore
[new]auth/python/flask/auth/tools/generate_secret_key

DESCRIPTION
Generate and store cryptographically secure flask secret keys

Add a tool to generate cryptographically secure secret keys
Add scripts to create a secret table in the database,
insert the generated secret keys into the database and
finally extract the secret key from the database.

For this application store the secret key in the
environment variable AUTH_APP_FLASK_SECRET_KEY.

Replace the dummy secret key assignment with code
that gets the secret key from the environment variable.
It should raise a RuntimeError (which aborts the application
before it can run) should the secret key not be present.

Outside of this commit, a secret flask key was generated
using the genrate_secret_key script and stored in the
databse using the database access scripts developed in
this commit.

The procedure to run this server now is first set the
environment variable with the secret key and then
run the application.