[issue002a1] Return HTTP 'Forbidden' instead of 'Not Found' for CSRF abort
FEATURE(s):
issue002a-auth-python:
  Create the auth application using python
  issue002a1-auth-python-flask:
    Create the auth application using python and flask

FILE(s):
[modified]auth/python/flask/auth/auth/auth.py
  Return HTTP 'Forbidden' instead of 'Not Found' for CSRF abort

DESCRIPTION
Return HTTP 'Forbidden' instead of 'Not Found' for CSRF abort.

The original logic for 'Not Found' was to avoid tipping off
a fraudulent signin attempt. But it seemed like it was better
to simply return Forbidden which is the correct RESTful response.
This might even discourage experimentation by an attacker.